Move shared method to py file
